refactor: moving common row to types folder
