+++ ui/src/lib/features/k8s/namespaces/store.ts
@@ -2,13 +2,9 @@
 // SPDX-FileCopyrightText: 2024-Present The UDS Authors
 
 import type { V1Namespace as Resource } from '@kubernetes/client-node'
-import {
-  ResourceStore,
-  type ColumnWrapper,
-  type CommonRow,
-  type ResourceStoreInterface,
-  type ResourceWithTable,
-} from '../store'
+
+import { ResourceStore, type ColumnWrapper, type ResourceStoreInterface, type ResourceWithTable } from '../store'
+import { type CommonRow } from '../types'
 
 export interface Row extends CommonRow {
   status: string
