+++ ui/src/lib/features/k8s/types.ts
@@ -1,0 +1,9 @@
+export interface CommonRow {
+  name: string
+  namespace?: string
+  creationTimestamp: Date
+  age?: {
+    sort: number
+    text: string
+  }
+}
